+++ BCC/userBDOS/LIB/WIZ5500.C
@@ -174,7 +174,7 @@ void wizWrite(word addr, word cb, char* buf, word len)
   WizSpiBeginTransfer();
 
   // Send address
-  word addrMSB = addr >> 8;
+  word addrMSB = (unsigned) addr >> 8;
   WizSpiTransfer(addrMSB); //msByte
   WizSpiTransfer(addr); //lsByte
 
@@ -198,7 +198,7 @@ word wizWriteSingle(word addr, word cb, word data)
   WizSpiBeginTransfer();
 
   // Send address
-  word addrMSB = addr >> 8;
+  word addrMSB = (unsigned) addr >> 8;
   WizSpiTransfer(addrMSB); //msByte
   WizSpiTransfer(addr); //lsByte
 
@@ -220,7 +220,7 @@ void wizWriteDouble(word addr, word cb, word data)
   WizSpiBeginTransfer();
 
   // Send address
-  word addrMSB = addr >> 8;
+  word addrMSB = (unsigned) addr >> 8;
   WizSpiTransfer(addrMSB); //msByte
   WizSpiTransfer(addr); //lsByte
 
@@ -228,7 +228,7 @@ void wizWriteDouble(word addr, word cb, word data)
   WizSpiTransfer(cb);
 
   // Send data
-  word dataMSB = data >> 8;
+  word dataMSB = (unsigned) data >> 8;
   WizSpiTransfer(dataMSB);
   WizSpiTransfer(data);
 
@@ -241,7 +241,7 @@ void wizRead(word addr, word cb, char* buf, word len)
   WizSpiBeginTransfer();
 
   // Send address
-  word addrMSB = addr >> 8;
+  word addrMSB = (unsigned) addr >> 8;
   WizSpiTransfer(addrMSB); //msByte
   WizSpiTransfer(addr); //lsByte
 
@@ -264,7 +264,7 @@ word wizReadSingle(word addr, word cb)
   WizSpiBeginTransfer();
 
   // Send address
-  word addrMSB = addr >> 8;
+  word addrMSB = (unsigned) addr >> 8;
   WizSpiTransfer(addrMSB); //msByte
   WizSpiTransfer(addr); //lsByte
 
@@ -286,7 +286,7 @@ word wizReadDouble(word addr, word cb)
   WizSpiBeginTransfer();
 
   // Send address
-  word addrMSB = addr >> 8;
+  word addrMSB = (unsigned) addr >> 8;
   WizSpiTransfer(addrMSB); //msByte
   WizSpiTransfer(addr); //lsByte
 
@@ -341,7 +341,7 @@ word wizGetSockReg8(word s, word addr){
   WizSpiBeginTransfer();
 
   // Send address
-  word addrMSB = addr >> 8;
+  word addrMSB = (unsigned) addr >> 8;
   WizSpiTransfer(addrMSB); //msByte
   WizSpiTransfer(addr); //lsByte
 
@@ -367,7 +367,7 @@ void wizSetSockReg16(word s, word addr, word val)
   WizSpiTransfer(0); //msByte
   WizSpiTransfer(addr); //lsByte
   WizSpiTransfer(WIZNET_WRITE_SnR + (s << 5));
-  word valMSB = val >> 8;
+  word valMSB = (unsigned) val >> 8;
   WizSpiTransfer(valMSB);
   WizSpiTransfer(val);
   WizSpiEndTransfer();
@@ -380,7 +380,7 @@ word wizGetSockReg16(word s, word addr)
   WizSpiBeginTransfer();
 
   // Send address
-  word addrMSB = addr >> 8;
+  word addrMSB = (unsigned) addr >> 8;
   WizSpiTransfer(addrMSB); //msByte
   WizSpiTransfer(addr); //lsByte
